Refactor code and create test scenario

+++ src/main/java/pl/ztbd/project/oracle/service/FlashcardsService.java
@@ -10,10 +10,12 @@ import pl.ztbd.project.api.dto.response.ResolveResponse;
 import pl.ztbd.project.oracle.entity.FlashcardEntity;
 import pl.ztbd.project.oracle.entity.FlashcardPageEntity;
 import pl.ztbd.project.oracle.entity.ResolvedPageEntity;
+import pl.ztbd.project.oracle.entity.UserEntity;
 import pl.ztbd.project.oracle.repository.FlashcardPageRepository;
 import pl.ztbd.project.oracle.repository.FlashcardRepository;
 import pl.ztbd.project.oracle.repository.ResolvedPageRepository;
 import pl.ztbd.project.oracle.repository.UserRepository;
+import pl.ztbd.project.security.JwtService;
 
 import java.util.List;
 
@@ -24,22 +26,33 @@ public class FlashcardsService implements FlashcardsAPI<Long> {
     private final FlashcardRepository flashcardRepository;
     private final FlashcardPageRepository flashcardPageRepository;
     private final ResolvedPageRepository resolvedPageRepository;
-
+    private final JwtService jwtService;
 
     @Override
-    public boolean addFlashcard(AddFlashcardRequest<Long> addFlashcardRequest) {
+    public Long addFlashcard(AddFlashcardRequest<Long> addFlashcardRequest) {
+        UserEntity userEntity = authenticateUser(addFlashcardRequest.token());
+        if (userEntity == null) {
+            return null;
+        }
+
         FlashcardEntity flashcardEntity = flashcardRepository.save(new FlashcardEntity(addFlashcardRequest.name(), addFlashcardRequest.description()));
         List<FlashcardPageEntity> flashcardPageEntities = addFlashcardRequest.pages()
                 .stream()
                 .map(page -> new FlashcardPageEntity(flashcardEntity.getId(), page.question(), page.answer()))
                 .toList();
         flashcardPageRepository.saveAll(flashcardPageEntities);
-        return true;
+        return flashcardEntity.getId();
     }
 
     @Override
-    public boolean addFlashcardPages(List<AddPageRequest<Long>> addPageRequest) {
+    public boolean addFlashcardPages(AddPagesRequest<Long> addPageRequest) {
+        UserEntity userEntity = authenticateUser(addPageRequest.token());
+        if (userEntity == null) {
+            return false;
+        }
+
         List<FlashcardPageEntity> flashcardPageEntities = addPageRequest
+                .addPageRequestList()
                 .stream()
                 .map(page -> new FlashcardPageEntity(page.flashcardId(), page.question(), page.answer()))
                 .toList();
@@ -49,18 +62,34 @@ public class FlashcardsService implements FlashcardsAPI<Long> {
 
     @Override
     public boolean removeFlashcard(RemoveFlashcardRequest<Long> removeFlashcardRequest) {
+        UserEntity userEntity = authenticateUser(removeFlashcardRequest.token());
+        if (userEntity == null) {
+            return false;
+        }
+
         flashcardRepository.deleteById(removeFlashcardRequest.flashcardId());
         return true;
     }
 
     @Override
     public boolean removeFlashcardPage(RemoveFlashcardPageRequest<Long> removeFlashcardPage) {
+        UserEntity userEntity = authenticateUser(removeFlashcardPage.token());
+        if (userEntity == null) {
+            return false;
+        }
+
         flashcardPageRepository.deleteById(removeFlashcardPage.pageId());
         return true;
     }
 
     @Override
     public boolean modifyFlashcard(ModifyFlashcardRequest<Long> modifyFlashcardRequest) {
+        UserEntity userEntity = authenticateUser(modifyFlashcardRequest.token());
+        if (userEntity == null) {
+            return false;
+        }
+
+
         return flashcardRepository.findById(modifyFlashcardRequest.flashcardId())
                 .map(flashcardEntity -> {
                     flashcardEntity.setDescription(modifyFlashcardRequest.description());
@@ -74,6 +103,11 @@ public class FlashcardsService implements FlashcardsAPI<Long> {
 
     @Override
     public boolean modifyFlashcardPage(ModifyFlashcardPageRequest<Long> modifyFlashcardPage) {
+        UserEntity userEntity = authenticateUser(modifyFlashcardPage.token());
+        if (userEntity == null) {
+            return false;
+        }
+
         return flashcardPageRepository.findById(modifyFlashcardPage.pageId())
                 .map(flashcardPageEntity -> {
                     flashcardPageEntity.setQuestion(modifyFlashcardPage.question());
@@ -86,6 +120,11 @@ public class FlashcardsService implements FlashcardsAPI<Long> {
 
     @Override
     public List<GetFlashcardsResponse<Long>> getFlashcards(GetFlashcardsRequest<Long> getFlashcardsRequest) {
+        UserEntity userEntity = authenticateUser(getFlashcardsRequest.token());
+        if (userEntity == null) {
+            return null;
+        }
+
         return flashcardRepository.findAll()
                 .stream()
                 .map(flashcardEntity -> new GetFlashcardsResponse<Long>(flashcardEntity.getId(), flashcardEntity.getName(), flashcardEntity.getDescription()))
@@ -94,6 +133,11 @@ public class FlashcardsService implements FlashcardsAPI<Long> {
 
     @Override
     public List<GetPageResponse<Long>> getPages(GetPageRequest<Long> getPageRequest) {
+        UserEntity userEntity = authenticateUser(getPageRequest.token());
+        if (userEntity == null) {
+            return null;
+        }
+
         return flashcardPageRepository.findAllByFlashcardId(getPageRequest.flashcardId())
                 .stream()
                 .map(flashcardPageEntity -> new GetPageResponse<Long>(flashcardPageEntity.getId(), flashcardPageEntity.getQuestion()))
@@ -102,14 +146,25 @@ public class FlashcardsService implements FlashcardsAPI<Long> {
 
     @Override
     public ResolveResponse resolve(ResolveRequest<Long> resolveRequest) {
+        UserEntity userEntity = authenticateUser(resolveRequest.token());
+        if (userEntity == null) {
+            return null;
+        }
+
         return flashcardPageRepository.findById(resolveRequest.pageId())
                 .map(flashcardPageEntity -> {
+
                     boolean isCorrect = flashcardPageEntity.getAnswer().equals(resolveRequest.answer());
-                    resolvedPageRepository.save(new ResolvedPageEntity(resolveRequest.userId(), resolveRequest.pageId(), resolveRequest.answer(), isCorrect));
+                    resolvedPageRepository.save(new ResolvedPageEntity(userEntity.getId(), resolveRequest.pageId(), resolveRequest.answer(), isCorrect));
                     return new ResolveResponse(flashcardPageEntity.getAnswer(), resolveRequest.answer(), isCorrect);
                 })
                 .orElse(null);
 
 
+    }
+
+    private UserEntity authenticateUser(String token) {
+        String email = jwtService.extractEmail(token);
+        return userRepository.findByEmail(email).orElse(null);
     }
 }
